+++ src/products/products.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model, FilterQuery } from 'mongoose';
+import { Model, FilterQuery, Types } from 'mongoose';
+import { InvalidObjectId } from 'src/exceptions/invalidObjectId.exception';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { ProductDocument } from './schemas/product.schema';
@@ -20,6 +21,9 @@ export class ProductsService {
   }
 
   findOne(id: string) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new InvalidObjectId(id);
+    }
     return this.productModel.findById(id);
   }
 
@@ -28,14 +32,23 @@ export class ProductsService {
   }
 
   update(id: string, updateProductDto: UpdateProductDto) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new InvalidObjectId(id);
+    }
     return this.productModel.findByIdAndUpdate(id, updateProductDto);
   }
 
   updateImage(id: string, image: string) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new InvalidObjectId(id);
+    }
     return this.productModel.findByIdAndUpdate(id, { image });
   }
 
   remove(id: string) {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new InvalidObjectId(id);
+    }
     return this.productModel.findByIdAndDelete(id);
   }
 }
